requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bleach==6.2.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorful==0.5.6
distlib==0.3.9
docopt==0.6.2
Faker==35.2.0
fastapi==0.115.8
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
graphframes==0.6
grpcio==1.70.0
h11==0.14.0
hdfs==2.7.3
httptools==0.6.4
idna==3.10
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
nose==1.3.7
numpy==2.2.2
opencensus==0.11.4
opencensus-context==0.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
py-spy==0.4.0
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
ray==2.41.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sparkmeasure==0.24.0
starlette==0.45.3
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
